Move geoip endpoint to nettools
